First "ok" version.
Rebased all previous commits into this one to cleanup the history
a bit.

Todo:

Split the index.js into components that can be loaded using the
module aggregation. This will simplify a bit and allow to make changes
to individual files instead of a big growing file.

There are currently 5 components Form / Select / Input / Output / Tags

The form works as a central data store that can be accessed by fields
each fields has access to everything in the top view.

Lit Element components do consume child elements that are later added
into the ShadowDom. I don't want to use slots on purpose. It could be
possible to use Slots instead but I prefer not to use them for some
specific reasons.

Slots are limited in the sense that they can only be used once. So if
you have a repeated list. The slot is of no use, while being able to
use html markup is useful. If you do not use slots, then you end up
with elements as children and in the shadow dom. The children elements
are never shown to the user anyway.. Maintaining a list of child element
can make rendering faster as the render function can simply
append back html elements without triggering a render to children.

So you can have ComponentA (render) -> ComponentB(render) ->
ComponentsC(no render)

If you have a list of 1000 children element but only the parent changed.
It will save you a lot of time and not using slots makes just the rest
of the code more consistent.